Fix the golang.sh.

+++ golang.sh
@@ -5,9 +5,15 @@
 
 URL="https://golang.org/dl/"
 ARCH=$(dpkg --print-architecture)
-LATEST=$(wget "$URL" -O - 2>/dev/null | grep Archive -B 1 | grep "linux-$ARCH" | head -n1 | sed -ne 's/.*\(http[^"]*\).*/\1/p')
+LINE=$(wget "$URL" -O - 2>/dev/null | grep Archive -B 1 | grep "linux-$ARCH" | head -n1)
+LATEST=$(echo "$LINE" | grep -Po '(?<=href=")[^"]*')
 FILE=$(basename "$LATEST")
 GOBIN="$HOME/.local/share/go/bin/go"
+
+if [ -z "$FILE" ]; then
+    echo "File not found."
+    exit 1
+fi
 
 # Check if there is an older golang.
 if [ -f "$GOBIN" ]; then
@@ -20,11 +26,15 @@ if [ -f "$GOBIN" ]; then
 fi
 
 SHA256=$(wget "$URL" -O - 2>/dev/null | grep "$LATEST" -A 5 | sed -ne 's/.*<tt>\([^<]*\)<\/tt>.*/\1/p')
+if [ -z "$SHA256" ]; then
+    echo "SHA256 not found."
+    exit 1
+fi
 echo "$SHA256" "$FILE"
 
 while [ ! -f "/tmp/$FILE" ] || [ "$(sha256sum "/tmp/$FILE" | awk '{print $1}')" != "$SHA256" ]; do
     rm -f "/tmp/$FILE"
-    wget -q "$LATEST" -O "/tmp/$FILE"
+    wget -q "https://golang.org$LATEST" -O "/tmp/$FILE"
 done
 
 cd ~/.local/share || exit
